Revert "[fix]Java does not have an unsigned integer type, so it will interpret negative numbers."
This reverts commit 8339aeec6e635d4b2dc21cad497764b01495dbf4.

revert commit;
fix this bug on server(The loss is within the highest acceptable range, but it has no impact and is sufficient for this project.)

+++ client/src/main/java/com/chencraft/ntu/model/FieldDefn.java
@@ -6,11 +6,11 @@ import lombok.Getter;
 public enum FieldDefn {
     NAME("name", String.class),
     PASSWORD("password", String.class),
-    ACCOUNT_NO("accountNumber", Long.class),
+    ACCOUNT_NO("accountNumber", Integer.class),
     CURRENCY("currency", Currency.class),
     INITIAL_BALANCE("initialBalance", Double.class),
     MONITOR_INTERVAL("monitorInterval", Integer.class),
-    DEST_ACCOUNT_NO("destAccountNumber", Long.class),
+    DEST_ACCOUNT_NO("destAccountNumber", Integer.class),
     AMOUNT("amount", Double.class);
 
     private final String fieldName;
